+++ offer-manager-domain/src/main/groovy/life/qbic/business/offers/content/CreateOfferContent.groovy
@@ -11,17 +11,7 @@ import life.qbic.datamodel.dtos.business.AffiliationCategory
 import life.qbic.datamodel.dtos.business.Offer
 import life.qbic.datamodel.dtos.business.OfferId
 import life.qbic.datamodel.dtos.business.ProductItem
-import life.qbic.datamodel.dtos.business.services.DataStorage
-import life.qbic.datamodel.dtos.business.services.MetabolomicAnalysis
-import life.qbic.datamodel.dtos.business.services.PrimaryAnalysis
-import life.qbic.datamodel.dtos.business.services.Product
-import life.qbic.datamodel.dtos.business.services.ProjectManagement
-import life.qbic.datamodel.dtos.business.services.ProteomicAnalysis
-import life.qbic.datamodel.dtos.business.services.SecondaryAnalysis
-import life.qbic.datamodel.dtos.business.services.Sequencing
-
-import java.text.DateFormat
-import java.time.Instant
+import life.qbic.datamodel.dtos.business.services.*
 
 /**
  * <h1>Creates the content for an offer export</h1>
@@ -161,7 +151,13 @@ class CreateOfferContent implements CreateOfferContentInput, FetchOfferOutput{
      * @return the discount percentage based on quantity discount and item total cost
      */
     private double calculateDiscountPercentage(ProductItem productItem) {
-        BigDecimal result = 100.0.toBigDecimal() * productItem.quantityDiscount.toBigDecimal() / productItem.totalPrice.toBigDecimal()
+        BigDecimal totalPrice = productItem.totalPrice.toBigDecimal()
+        if (totalPrice.compareTo(BigDecimal.ZERO) == 0) {
+            //avoid division by 0
+            //if a product has a total price of 0 set discount percentage to 0
+            return 0.doubleValue()
+        }
+        BigDecimal result = 100.0.toBigDecimal() * productItem.quantityDiscount.toBigDecimal() / totalPrice
         return result.doubleValue()
     }
     
@@ -172,10 +168,15 @@ class CreateOfferContent implements CreateOfferContentInput, FetchOfferOutput{
      * @return the discount per unit, if applicable, 0 otherwise
      */
     private double calculateDiscountPerUnit(ProductItem productItem) {
-        BigDecimal result = productItem.quantityDiscount.toBigDecimal() / productItem.quantity.toBigDecimal()
+        BigDecimal quantity = productItem.quantity.toBigDecimal()
+        if (quantity.compareTo(BigDecimal.ZERO) == 0) {
+            //avoid division by 0
+            //if a a productItem has no quantity set discount to 0
+            return 0.doubleValue()
+        }
+        BigDecimal result = productItem.quantityDiscount.toBigDecimal() / quantity
         return result.doubleValue()
     }
-
 
     /**
      * Adds the product items to the respective product group list