+++ CHANGELOG.rst
@@ -4,6 +4,50 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+1.2.0 (2021-09-17)
+------------------
+
+**Added**
+
+**Fixed**
+
+**Dependencies**
+
+**Deprecated**
+
+1.1.2 (2021-09-17)
+------------------
+
+**Added**
+
+**Fixed**
+
+* Catch exception during discount generation for negative product prices (`#808 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/808>`_)
+
+**Dependencies**
+
+**Deprecated**
+
+1.1.1 (2021-09-14)
+------------------
+
+**Added**
+
+**Fixed**
+
+* Avoid page breaks in footer content (`#798 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/798>`_)
+
+* Disable buttons after deselecting offers (`#799 <https://github.com/qbicsoftware/offer-manager-2-portlet/issues/799>`_)
+
+* Avoid division by zero during offer content creation (`#806 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/806>`_)
+
+**Dependencies**
+
+* Bump data model lib 2.12.0 -> 2.12.1 (includes important bug fixes)
+
+**Deprecated**
+
+
 1.1.0 (2021-08-17)
 ------------------
 
@@ -26,6 +70,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Add discount unit and percentage to discount items, specify the discount item description (`#755 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/755>`_)
 
 * Adds the total price before VAT is applied to an offer (`#754 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/754>`_)
+
+* Add new offer layout and css  (`#768 <https://github.com/qbicsoftware/offer-manager-2-portlet/pull/768>`_)
 
 **Fixed**